+++ lib/lightning/workorders/workorder.ex
@@ -64,5 +64,6 @@ defmodule Lightning.WorkOrder do
     changeset
     |> assoc_constraint(:workflow)
     |> assoc_constraint(:snapshot)
+    |> foreign_key_constraint(:snapshot_id)
   end
 end