+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to
   [#1822](https://github.com/OpenFn/lightning/issues/1822)
 - Fix sample workflow from init_project_for_new_user
   [#2016](https://github.com/OpenFn/lightning/issues/2016)
+- Benchmarking script that simulates data from a cold chain.
+  [#1993](https://github.com/OpenFn/lightning/issues/1993)
 
 ### Changed